Move train and evaluate to a separate file to remove code duplication
